pyproject.toml
[tool.poetry]
name = "slack-agent"
version = "0.1.0"
description = "LLM-powered slack agent for me."
authors = ["h-terao <h-terao@ymail.ne.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13.1"
ruff = "^0.8.2"
slack-bolt = "^1.21.3"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[tool.ruff]
line-length = 120
select = ["E4", "E7", "E9", "F"]
ignore = ["E731"]

[tool.ruff.lint.pydocstyle]
convention = "google"